Version 1.0 Push
Fixed some display errors. Updated TODO and README. Cleaned up code a
little bit.
This will be the final release. I'm content with how this project feels
in the current state. There are some more features that would be good,
but as this is a learning project so they will be going towards my next
3DS project.